Fix the delegates table definition, add attributes discovery tests
 * change the column type for delegate block_level
 * add discovery tests for attributes of: bakers, ballots, proposals,
   delegates

+++ doc/conseil.sql
@@ -198,7 +198,7 @@ CREATE TABLE public.delegates (
     delegated_balance numeric,
     deactivated boolean NOT NULL,
     grace_period integer NOT NULL,
-    block_level numeric DEFAULT '-1'::integer NOT NULL
+    block_level integer DEFAULT '-1'::integer NOT NULL
 );
 
 CREATE TABLE public.delegated_contracts (
